+++ R/introjs.R
@@ -1,17 +1,17 @@
 # Copyright (C) 2016 Carl Ganz
 #
 # This file is part of rintrojs.
-# 
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
 # published by the Free Software Foundation, either version 3 of the
 # License, or (at your option) any later version.
-# 
+#
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
-# 
+#
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
@@ -22,9 +22,9 @@
 #' @importFrom jsonlite toJSON
 #' @param session the Shiny session object (from the server function of the Shiny app)
 #' @param options List of options to be passed to intro.js
-#' @param events List of text that are the body of a Javascript function
+#' @param events List of text that are the body of a Javascript function. Must wrap in [I()]
 #' @note For documentation on intro.js options and events, see \url{https://github.com/usablica/intro.js/wiki/Documentation}.
-#' @seealso \code{\link{introjsUI}} \code{\link{introBox}}
+#' @seealso [introjsUI()] [introBox()]
 #' @examples
 #' \dontrun{
 #' library(rintrojs)
@@ -45,9 +45,9 @@
 #'     )
 #'   )))
 #' server <- shinyServer(function(input, output, session) {
-#' 
+#'
 #'  hintjs(session, options = list("hintButtonLabel"="That was a hint"))
-#' 
+#'
 #'   output$mtcars <- renderTable({
 #'     head(mtcars)
 #'   })
@@ -60,33 +60,41 @@
 #' }
 #' @export
 
-introjs <- function(session, options = list(), events = list()) {
+introjs <- function(session,
+                    options = list(),
+                    events = list()) {
   options <- list(options = options, events = events)
-  session$sendCustomMessage(type = "introjs", message = jsonlite::toJSON(options))
+  session$sendCustomMessage(type = "introjs",
+                            message = jsonlite::toJSON(options,
+                                                       auto_unbox = TRUE))
   
 }
 
 #' @rdname introjs
 #' @export
 
-hintjs <- function(session, options= list(), events = list()) {
+hintjs <- function(session,
+                   options = list(),
+                   events = list()) {
   options <- list(options = options, events = events)
   session$sendCustomMessage(type = "hintjs", message = jsonlite::toJSON(options))
   
 }
 
 #' Set up Shiny app to use intro.js
-#' 
+#'
 #' This function must be called from a Shiny app's UI in order
 #' to use the package.
-#' 
+#'
 #' @param includeOnly Only include intro.js files. For users who will write their own javascript
+#' @param cdn Indicate whether to include intro.js files from CDN
+#' @param version Specify intro.js version to use from cdn
 #' @export
-#' @examples 
+#' @examples
 #' \dontrun{
 #' library(rintrojs)
 #' library(shiny)
-#' 
+#'
 #' shinyApp(
 #' ui = fluidPage(
 #'   introjsUI(), # must include in UI
@@ -102,40 +110,43 @@ hintjs <- function(session, options= list(), events = list()) {
 #' )
 #' }
 
-introjsUI <- function(includeOnly = FALSE) {
-  
-  shiny::tags$head(shiny::singleton(shiny::tagList(
-    shiny::includeScript(
-      system.file("javascript/introjs/intro.min.js", package = "rintrojs")
-    ),
-    shiny::includeCSS(
-      system.file(
-        "javascript/introjs/introjs.min.css",
-        package = "rintrojs"
-      )
-    ),
-    if (!includeOnly) {
-      shiny::includeScript(
-        system.file(
-          "javascript/rintro.js", package = "rintrojs"
-        )
-      )
-    }
-  )
-  )
-  )
+introjsUI <- function(includeOnly = FALSE, cdn = FALSE, version = "2.5.0") {
+  if (!missing(version) && !cdn) {
+    warning("version parameter is ignored when cdn = FALSE")
+  }
+  shiny::tags$head(shiny::singleton(
+    shiny::tagList(
+      shiny::includeScript(if (cdn) {
+        paste0("https://cdn.jsdelivr.net/intro.js/",
+               version, "/intro.min.js")
+      } else {
+        system.file("javascript/introjs/intro.min.js", package = "rintrojs")
+      }),
+      shiny::includeCSS(if (cdn) {
+        paste0("https://cdn.jsdelivr.net/intro.js/",
+               version, "/introjs.min.css")
+      } else {
+        system.file("javascript/introjs/introjs.min.css",
+                    package = "rintrojs")
+      }),
+      if (!includeOnly) {
+        shiny::includeScript(system.file("javascript/rintro.js", package = "rintrojs"))
+      }
+    )
+  ))
 }
 
 #' Generate intro elements in UI
 #'
-#' Wrap \code{introBox} around elements you want to include in introduction.
+#' Wrap `introBox` around elements you want to include in introduction.
 #' Use data.step to order the boxes and data.intro to specify the comment in the introduction
 #'
 #' @param ... Elements in introduction element
 #' @param data.step a number indicating its spot in the order in the intro
 #' @param data.intro text for introduction
 #' @param data.hint text for clickable hints
-#' @seealso \code{\link{introjsUI}} \code{\link{introjs}}
+#' @param data.position position of intro
+#' @seealso [introjsUI()] [introjs()]
 #' @examples
 #' \dontrun{
 #' library(rintrojs)
@@ -166,15 +177,38 @@ introjsUI <- function(includeOnly = FALSE) {
 #' }
 #' @export
 
-introBox <- function(... ,data.step, data.intro, data.hint) {
-  stopifnot(!((!missing(data.step) & missing(data.intro)) | (missing(data.step) & !missing(data.intro))))
-  data <- match.call(expand.dots = TRUE)
-  n <- length(list(...)) + 1
-  names(data)[-seq_len(n)] <- gsub("\\.", "-", names(data)[-seq_len(n)])
-  data[[1]] <- quote(shiny::tags$div)
-  # http://stackoverflow.com/a/40180906/4564432
-  shiny::singleton(eval.parent(data))
-  
-}
-
-
+introBox <-
+  function(... ,
+           data.step,
+           data.intro,
+           data.hint,
+           data.position = c(
+             "bottom",
+             "auto",
+             "top",
+             "left",
+             "right",
+             "bottom",
+             "bottom-left_aligned",
+             "bottom-middle-aligned",
+             "bottom-right-aligned",
+             "auto"
+           )) {
+    stopifnot(!((
+      !missing(data.step) &
+        missing(data.intro)
+    ) | (
+      missing(data.step) & !missing(data.intro)
+    )))
+    
+    data.position <- match.arg(data.position)
+    
+    data <- match.call(expand.dots = TRUE)
+    n <- length(list(...)) + 1
+    names(data)[-seq_len(n)] <-
+      gsub("\\.", "-", names(data)[-seq_len(n)])
+    data[[1]] <- quote(shiny::tags$div)
+    # http://stackoverflow.com/a/40180906/4564432
+    shiny::singleton(eval.parent(data))
+    
+  }